+++ feather/game.h
@@ -1,8 +1,8 @@
 #include <stdint.h>
-extern int p1;
-extern int p2;
-extern int bx;
-extern int by;
+extern uint8_t p1;
+extern uint8_t p2;
+extern uint8_t bx;
+extern uint8_t by;
 
 extern uint8_t button_status;
